Add conditional check to adding column in migration file

+++ bundles/CoreBundle/Migrations/Version20230223101848.php
@@ -24,11 +24,15 @@ final class Version20230223101848 extends AbstractMigration
 {
     public function up(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE `users` ADD COLUMN `provider` varchar(255) DEFAULT NULL;');
+        if (!$schema->getTable('users')->hasColumn('provider')) {
+            $this->addSql('ALTER TABLE `users` ADD COLUMN `provider` varchar(255) DEFAULT NULL;');
+        }
     }
 
     public function down(Schema $schema): void
     {
-        $this->addSql('ALTER TABLE `users` DROP COLUMN `provider`;');
+        if ($schema->getTable('users')->hasColumn('provider')) {
+            $this->addSql('ALTER TABLE `users` DROP COLUMN `provider`;');
+        }
     }
 }
